(core) persist logins across docker restarts by default
Summary:
Unless redis is configured, Grist will store information in a
grist-sessions.db file. For grist-core run using docker, this
file is not by default in the directory the documentation suggests
the user persist, and so is lost when Grist is updated. This
is an unexpected annoyance for users, and the configuration fix
is awkward to explain, so this diff changes the default.

Test Plan:
rebuilt image manually and verified it operated as
expected

Reviewers: georgegevoian

Differential Revision: https://phab.getgrist.com/D3217