back off from making gvisor sandboxing default in docker image
It looks like making gvisor sandboxing the default in our docker
image is causing people trouble, so this backs off from that
change. We retain gvisor's runsc executable in the image so that
turning on sandboxing is just an environment variable setting away.

Lack of sandboxing is not good for users opening untrusted documents,
so it would be good to be aggressive about turning it on, or
communicating about it, so there's follow-up work needed.
In the meantime I've updated the documentation about it somewhat.